requirements.txt
PyQt5
requests
psycopg2-binary
nose2